updated so that each slot is a single component instance, also created slot component
