+++ packages/app/src/components/StoryFactory/index.tsx
@@ -4,3 +4,4 @@ export * from "./FabricaCard";
 export * from "./SF_SlotMachineGame";
 export * from "./StoryFactoryButton";
 export * from "./StoriesCardNoRating";
+export * from "./SF_Slot";
